app/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
pydantic>=2.0.0
PyJWT>=2.0.0
httpx>=0.27.0
google-genai>=1.0.0
